scripts/utils/generation-favicons/requirements.txt
click==8.1.3
Pillow==9.5.0
pkg_resources==0.0.0
typer==0.7.0
